dashboard/requirements.txt
google-cloud-bigquery>=3.40.1
python-dotenv>=1.2.1
streamlit>=1.54.0
pandas>=2.3.3
plotly>=6.5.2
db-dtypes>=1.5.0
